+++ spec/run.js
@@ -3,6 +3,6 @@ var jasmine = new (require('jasmine'))(),
     path = require('path')
 
 jasmine.configureDefaultReporter({print: function () {}})
-jasmine.loadConfigFile(path.join(__dirname, 'jasmine.json'))
+jasmine.loadConfigFile(path.join(__dirname, 'support', 'jasmine.json'))
 jasmine.addReporter(new SpecReporter())
 jasmine.execute(process.argv.slice(2))